+++ media/main.js
@@ -45,6 +45,7 @@ $(document).ready(function () {
       )
       .replaceAll("/@", "/");
   };
+
   const addPRsToDoc = (prs) => {
     $("#ghHolder").append(
       "<button class='run-watermelon'>Run Watermelon</button><br/>"
@@ -54,6 +55,9 @@ $(document).ready(function () {
     );
     $(".run-watermelon").on("click", (event) => {
       sendMessage({ command: "run" });
+    });
+    $(".help-link").on("click", (event) => {
+      sendMessage({ command: "open-link", link: "https://app.slack.com" });
     });
     prs.forEach((pr, index) => {
       let mdComments = "";
@@ -87,7 +91,7 @@ $(document).ready(function () {
         <div>
           <div class="pr-owner">
             <p class="pr-poster" title="View this user on github">
-              Author: <a href="${pr.userLink}">${pr.user}</a>
+              <a class="pr-author-combo" href="${pr.userLink}"><img class='pr-author-img' src="${pr.userImage}" />${pr.user}</a>
             </p>
             <p class="pr-date">
               ${new Date(pr.created_at).toLocaleDateString("en-us", {
@@ -138,16 +142,18 @@ $(document).ready(function () {
       <p>Try running a new Watermelon query, please.</p>
     </div>
     `);
-    $("#ghHolder")
-      .append("<button class='run-watermelon'>Run Watermelon</button><br/>")
-      .on("click", (event) => {
-        sendMessage({ command: "run" });
-      });
-    $("#ghHolder")
-      .append(`<button class='help-link' >Get help from ${authorName}</button>`)
-      .on("click", (event) => {
-        sendMessage({ command: "open-link", link: "https://app.slack.com" });
-      });
+    $("#ghHolder").append(
+      "<button class='run-watermelon'>Run Watermelon</button><br/>"
+    );
+    $(".run-watermelon").on("click", (event) => {
+      sendMessage({ command: "run" });
+    });
+    $("#ghHolder").append(
+      `<button class='help-link' >Get help from ${authorName}</button>`
+    );
+    $(".help-link").on("click", (event) => {
+      sendMessage({ command: "open-link", link: "https://app.slack.com" });
+    });
 
     $("#ghHolder").append(
       "<p>Alternatively, you can <a href='https://github.com/watermelontools/wm-extension#commands'>run with our watermelon.start command</a></p>"