+++ CHANGELOG.md
@@ -10,8 +10,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Will now log the extension version to the console on startup
 - The get help button now shows the top committer name instead of "on Slack"
 - Fixed an error where the command would break even when results were fetched
+- Fixed some instances of the get help on Slack button not working
 - Will select whole file if no selection is made
 - Fixed bug that retrieved PRs from unrelated repositories
+- Added Connect-src to CSP
+- Added PR author image
 
 ## [1.1.3]
 - Fixed some instances of the "Get help on Slack" running search